backend/webapi/pyproject.toml
[tool.poetry]
name = "webapi"
version = "0.1.0"
description = ""
authors = ["slorello89 <steve.lorello@redis.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.109.0"
uvicorn = "^0.26.0"
python-dotenv = "^1.0.0"
semantic-kernel = ">=1.1.1"
pydantic = ">2"
requests = "^2.31.0"
python-multipart = "^0.0.6"
pypdf2 = "^3.0.1"
redis = "^5.0.1"
python-ulid = "^1.0.3"

[tool.poetry.scripts]
start = "webapi.main:start"


[tool.poetry.group.dev.dependencies]
setuptools = "^69.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
